+++ packages/ingest-mongodb-public/src/sources/prisma.ts
@@ -1,4 +1,4 @@
-import { makeMdOnGithubDataSource } from "mongodb-rag-core";
+import { makeMdOnGithubDataSource } from "mongodb-rag-core/dataSources";
 
 /**
   This is necessary for the Prisma source because the Prisma docs
