+++ packages/ingest-mongodb-public/src/sources/snooty/SnootyDataSource.ts
@@ -1,7 +1,7 @@
 import { createInterface } from "readline";
 import { Page, PageFormat, logger } from "mongodb-rag-core";
 import fetch from "node-fetch";
-import { DataSource, ProjectBase } from "mongodb-rag-core";
+import { DataSource, ProjectBase } from "mongodb-rag-core/dataSources";
 import { snootyAstToMd, getTitleFromSnootyAst } from "./snootyAstToMd";
 import {
   getTitleFromSnootyOpenApiSpecAst,
