+++ packages/ingest-mongodb-public/src/sources/wiredTiger.ts
@@ -2,7 +2,7 @@ import {
   HandleHtmlPageFuncOptions,
   handleHtmlDocument,
   makeGitDataSource,
-} from "mongodb-rag-core";
+} from "mongodb-rag-core/dataSources";
 
 const removeElements = (domDoc: Document) => [
   ...Array.from(domDoc.querySelectorAll("head")),
