+++ packages/mongodb-rag-core/src/index.ts
@@ -6,8 +6,8 @@ export * from "./contentStore";
 export * as ContentStore from "./contentStore";
 export * from "./conversations";
 export * as Conversations from "./conversations";
-export * from "./dataSources";
-export * as DataSources from "./dataSources";
+// export * from "./dataSources";
+// export * as DataSources from "./dataSources";
 export * from "./embed";
 export * as Embed from "./embed";
 export * from "./findContent";
